test: split prompt proof label policy coverage
