+++ test/clawsweeper.test.ts
@@ -11,7 +11,6 @@ import {
   statSync,
   writeFileSync,
 } from "node:fs";
-import { tmpdir } from "node:os";
 import { delimiter, join } from "node:path";
 import test from "node:test";
 
@@ -29,47 +28,24 @@ import {
   dashboardClosedAt,
   failedReviewRetryEligibilityForTest,
   fixedPullRequestFromCommitPullsForTest,
-  featureShowcaseLabelsForTest,
   formatRecentClosedRows,
   ghRetryKind,
   hotIntakeRecencyMs,
   isCodexReviewCommentBody,
   isInfrastructureFailedReviewForTest,
-  isGitHubLabelCapacityErrorForTest,
   isGitHubNotFoundError,
   isGitHubRequiresAuthenticationError,
   isLockedConversationCommentError,
-  isMissingGitHubLabelErrorForTest,
-  issueAdvisoryLabelsForTest,
   isProtectedItem,
-  labelJustificationsMarkdownForTest,
   itemNumbersArg,
   lockedConversationApplyReason,
   makeTreeReadOnlyForTest,
   parseGhJson,
   parseGhJsonLines,
   parseDecision,
-  mergeRiskLabelsForTest,
-  mergeRiskLabelSchemeForTest,
-  prRatingLabelsForTest,
-  prRatingLabelSchemeForTest,
-  prepareMediaProofArtifactsForTest,
-  proofVideoUrlsFromContextForTest,
-  prStatusLabelsForTest,
-  prStatusLabelSchemeForTest,
-  priorityLabelsForTest,
-  priorityLabelSchemeForTest,
-  proofNudgeCandidateRecordsForTest,
-  proofNudgeEligibilityForTest,
   protectedLabels,
-  realBehaviorProofMediaLabelsForTest,
-  realBehaviorProofSufficientLabelsForTest,
-  botProofCandidateRecordsForTest,
-  botProofEligibilityForTest,
   relatedGitHubIssueSearchQueryForTest,
   relatedTitleSearchTerms,
-  renderBotProofDecisionCommentForTest,
-  renderProofNudgeCommentForTest,
   renderReviewStartStatusComment,
   reviewArtifactDestination,
   reviewAutomationMarkersFromReport,
@@ -87,8 +63,6 @@ import {
   reviewPromptTemplate,
   runCodexForTest,
   lowerCodexReasoningEffort,
-  impactLabelsForTest,
-  impactLabelSchemeForTest,
   runtimeBudgetExceeded,
   safeOutputTail,
   redactInternalCodexModel,
@@ -101,7 +75,6 @@ import {
   shouldReviewItem,
   shouldRetryGh,
   shouldPlanItem,
-  telegramVisibleProofLabelsForTest,
   validateCloseDecision,
 } from "../dist/clawsweeper.js";
 import { checkConclusionForFrontMatter } from "../dist/commit-checks.js";
@@ -13269,2065 +13242,6 @@ test("root-cause report parsing defaults legacy and malformed reports safely", (
       }),
     ),
     valid,
-  );
-});
-
-test("review prompt routes PR likely owners through feature history", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /feature-history hunt/);
-  assert.match(prompt, /who introduced the feature/);
-  assert.match(prompt, /git log --follow -- <file>/);
-  assert.match(prompt, /do not list the PR author solely/);
-  assert.match(prompt, /not to the PR\s+author merely for writing the proposal/);
-  assert.match(prompt, /Do\s+not use `maintainer` as a likely-owner role/);
-  assert.match(prompt, /Do not include email\s+addresses in `likelyOwners`/);
-  assert.match(prompt, /use names without email addresses/);
-});
-
-test("review prompt describes concrete review metrics without vague examples", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /Always fill `reviewMetrics`/);
-  assert.match(prompt, /useful, concrete, maintainer-relevant/);
-  assert.match(prompt, /2 added, 1 changed, 0\s+removed/);
-  assert.match(prompt, /Do not use vague\s+labels or values/);
-  assert.doesNotMatch(prompt, /Risky change/);
-  assert.doesNotMatch(prompt, /Some changes/);
-  assert.doesNotMatch(prompt, /This seems risky/);
-});
-
-test("review prompt reads maintainer notes before PR diffs", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /\.agents\/maintainer-notes\//);
-  assert.match(prompt, /before reviewing the diff/);
-  assert.match(prompt, /Treat matching notes as maintainer decisions/);
-  assert.match(prompt, /do not publish raw internal note contents/);
-});
-
-test("review prompts treat target AGENTS as optional review policy", () => {
-  const itemPrompt = readFileSync("prompts/review-item.md", "utf8");
-  const commitPrompt = readFileSync("prompts/review-commit.md", "utf8");
-
-  for (const prompt of [itemPrompt, commitPrompt]) {
-    assert.match(
-      prompt,
-      /Before reviewing, read the target\s+repository's full `AGENTS\.md` file if present/,
-    );
-    assert.match(prompt, /Do not rely only on search\s+snippets/);
-    assert.match(
-      prompt,
-      /`head` output, local excerpts, partial line ranges, or truncated\s+copies/,
-    );
-    assert.match(prompt, /optional\s+repository-authored\s+review policy and review guidance/);
-    assert.match(
-      prompt,
-      /do not conflict with this prompt or higher-priority\s+system\/developer\s+instructions/,
-    );
-    assert.match(prompt, /existing repository\s+profiles and owner\/default fallback behavior/);
-    assert.match(prompt, /Use target `AGENTS\.md` policy as review input/);
-  }
-
-  assert.match(itemPrompt, /report it through `reviewFindings`/);
-  assert.match(
-    itemPrompt,
-    /route the\s+concern through the existing `risks`, `bestSolution`, `solutionAssessment`, or\s+`workReason` fields/,
-  );
-  assert.match(
-    commitPrompt,
-    /Report an AGENTS-policy conflict only when the commit creates a\s+concrete bug/,
-  );
-  assert.match(commitPrompt, /keep it out of `result: findings`/);
-});
-
-test("review prompt requires a dedicated securityReview section", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /Always summarize this pass in `securityReview`/);
-  assert.match(prompt, /Always fill `securityReview`/);
-  assert.match(prompt, /status: "needs_attention"/);
-});
-
-test("review prompt treats duplicated behavior as a P1 PR finding", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /dedicated solution-fit and upgrade-safety pass/);
-  assert.match(prompt, /current code, documented configuration, CLI flags, env vars/);
-  assert.match(prompt, /Search the codebase and docs for the existing capability/);
-  assert.match(prompt, /Treat duplicated behavior as a high-priority defect/);
-  assert.match(prompt, /add a P1 review finding unless the PR proves/);
-  assert.match(prompt, /maintenance drift, conflicting behavior,\s+or user confusion/);
-});
-
-test("review prompt treats plugin API changes as compatibility-sensitive P1 repair work", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /Treat plugin API surface changes as compatibility-sensitive/);
-  assert.match(prompt, /adds,\s+removes, renames, deprecates, changes behavior for/);
-  assert.match(prompt, /adds new similar\/parallel\s+calls to a plugin API/);
-  assert.match(prompt, /require explicit maintainer-visible discussion/);
-  assert.match(prompt, /Use\s+`merge-risk: 🚨 compatibility`/);
-  assert.match(prompt, /name the plugin API concern in `risks`/);
-  assert.match(prompt, /make\s+`mergeRiskOptions` spell out the maintainer choices or repair path/);
-  assert.match(prompt, /Prefer a\s+resolvable P1 review finding/);
-  assert.match(prompt, /preserving the existing API/);
-  assert.match(prompt, /removing the duplicate\/parallel call/);
-  assert.match(prompt, /clear deprecation path/);
-  assert.match(prompt, /focused\s+compatibility tests/);
-  assert.match(
-    prompt,
-    /Choose\s+`queue_fix_pr` for plugin API findings only when the\s+repair is concrete/,
-  );
-  assert.match(
-    prompt,
-    /Use\s+`manual_review` when the unresolved blocker is whether the new API should exist/,
-  );
-});
-
-test("review prompt requires upgrade and preference overwrite checks", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /Treat compatibility and user settings as merge-critical/);
-  assert.match(prompt, /override existing preferences, persisted config, provider choices/);
-  assert.match(
-    prompt,
-    /A new default must not change an existing user's stored\s+value during upgrade/,
-  );
-  assert.match(prompt, /Call out upgrade and settings breakage directly in `reviewFindings`/);
-  assert.match(prompt, /existing config\/preferences can be overwritten/);
-  assert.match(prompt, /preserving the existing\s+behavior as the default/);
-  assert.match(prompt, /explicit strict config option/);
-  assert.match(prompt, /default compatibility mode and the\s+opt-in strict mode/);
-  assert.match(prompt, /require evidence for both fresh-install behavior and upgrade\s+behavior/);
-  assert.match(prompt, /If upgrade behavior is ambiguous, mark the PR incorrect/);
-});
-
-test("review prompt treats stored data-model changes as compatibility-sensitive", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /Treat stored data-model changes as compatibility-sensitive/);
-  assert.match(prompt, /SQL\s+DDL or migrations/);
-  assert.match(prompt, /persistent cache schemas/);
-  assert.match(prompt, /Durable Object or hosted storage schemas/);
-  assert.match(prompt, /serialized JSON state written to disk/);
-  assert.match(prompt, /vector or embedding row identity\/query-compatibility metadata/);
-  assert.match(prompt, /doctor, repair, migration, or backfill code/);
-  assert.match(prompt, /pure query-only changes or non-semantic docs wording/);
-  assert.match(prompt, /migration or upgrade compatibility proof before any pass/);
-});
-
-test("review prompt requires real behavior proof for PR reviews", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /realBehaviorProof/);
-  assert.match(prompt, /Terminal screenshots|terminal screenshots/);
-  assert.match(prompt, /download\/open GitHub attachment links/);
-  assert.match(prompt, /generate stills or contact sheets from videos/);
-  assert.match(prompt, /compare the proof against the PR diff/);
-  assert.match(prompt, /Prefer asking for screenshots or videos/);
-  assert.match(prompt, /redact private information like IP addresses, API keys/);
-  assert.match(prompt, /screenshot-only proof sufficient/);
-  assert.match(prompt, /no visible console violation/);
-  assert.match(prompt, /scratch directory/);
-  assert.match(prompt, /@clawsweeper re-review/);
-  assert.match(
-    prompt,
-    /Unit tests, mocks, snapshots, lint, typechecks, and CI are supplemental only/,
-  );
-  assert.match(prompt, /do not request ClawSweeper repair markers/);
-});
-
-test("media proof preparation extracts browser-unplayable ffmpeg-decodeable video proof", () => {
-  const dir = mkdtempSync(join(tmpdir(), "clawsweeper-media-proof-"));
-  try {
-    const context = {
-      issue: {},
-      comments: [
-        {
-          body: [
-            "Chromium media error code 4 on this upload, but ffmpeg can decode it:",
-            "https://github.com/user/repo/releases/download/proof/Screen.Recording.mov",
-          ].join("\n"),
-        },
-      ],
-      timeline: [],
-    };
-    const calls: string[] = [];
-    const prepared = prepareMediaProofArtifactsForTest(context, dir, (command, args) => {
-      calls.push(`${command} ${args.join(" ")}`);
-      if (command === "curl") {
-        const outputIndex = args.indexOf("--output");
-        assert.notEqual(outputIndex, -1);
-        writeFileSync(String(args[outputIndex + 1]), "fake mov bytes");
-        return { status: 0, stdout: "", stderr: "" };
-      }
-      if (command === "ffprobe") {
-        return {
-          status: 0,
-          stdout: JSON.stringify({
-            format: { duration: "46.49" },
-            streams: [{ codec_name: "h264", width: 734, height: 1038 }],
-          }),
-          stderr: "",
-        };
-      }
-      if (command === "ffmpeg") {
-        const output = String(args.at(-1));
-        writeFileSync(output, "fake contact sheet");
-        return { status: 0, stdout: "", stderr: "" };
-      }
-      return { status: 1, stdout: "", stderr: `unexpected command: ${command}` };
-    });
-
-    assert.equal(prepared.artifacts.length, 1);
-    assert.equal(prepared.artifacts[0]?.status, "prepared");
-    assert.ok(prepared.manifestPath);
-    assert.ok(prepared.summaryPath);
-    assert.ok(prepared.artifacts[0]?.metadataPath);
-    assert.ok(prepared.artifacts[0]?.contactSheetPath);
-    assert.equal(existsSync(prepared.manifestPath), true);
-    assert.equal(existsSync(prepared.artifacts[0].metadataPath), true);
-    assert.equal(existsSync(prepared.artifacts[0].contactSheetPath), true);
-    assert.match(calls.join("\n"), /^curl /m);
-    assert.match(calls.join("\n"), /^ffprobe /m);
-    assert.match(calls.join("\n"), /^ffmpeg /m);
-  } finally {
-    rmSync(dir, { recursive: true, force: true });
-  }
-});
-
-test("runtime prompt tells Codex to inspect ffmpeg video artifacts before browser fallback", () => {
-  const context = {
-    issue: {},
-    comments: [{ body: "Proof: https://github.com/user/repo/releases/download/proof/demo.mov" }],
-    timeline: [],
-  };
-  const prompt = reviewPromptForTest(
-    item({ kind: "pull_request" }),
-    context,
-    { mainSha: "abc123", latestRelease: null },
-    "",
-    {
-      proofScratchDir: "/tmp/proof",
-      mediaProofManifestPath: "/tmp/proof/media-proof-manifest.json",
-      mediaProofSummary: "prepared: https://github.com/user/repo/releases/download/proof/demo.mov",
-    },
-  );
-
-  assert.deepEqual(proofVideoUrlsFromContextForTest(context), [
-    "https://github.com/user/repo/releases/download/proof/demo.mov",
-  ]);
-  assert.match(prompt, /preprocessed linked video proof with ffprobe\/ffmpeg/);
-  assert.match(prompt, /generated contact-sheet image paths before trying browser playback/);
-  assert.match(
-    prompt,
-    /Only fall back to browser playback after checking the prepared ffmpeg artifacts/,
-  );
-  assert.match(
-    prompt,
-    /If browser playback fails but ffprobe metadata and ffmpeg contact sheets are readable/,
-  );
-});
-
-test("review prompt keeps draft and protected workflow state out of PR rank", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /Rate PR evidence\s+and patch quality/);
-  assert.match(prompt, /weaker proof-or-patch quality signal/);
-  assert.match(
-    prompt,
-    /Do not lower `proofTier`, `patchTier`,\s+or `overallTier` solely because the PR is draft/,
-  );
-  assert.match(prompt, /has protected labels/);
-  assert.match(prompt, /not\s+automerge-eligible/);
-  assert.match(prompt, /workflow\s+state signals, not proof or patch quality defects/);
-});
-
-test("decision schema keeps draft and protected workflow state out of PR rank", () => {
-  const schema = JSON.parse(readFileSync("schema/clawsweeper-decision.schema.json", "utf8"));
-  const prRating = schema.properties.prRating;
-
-  assert.match(prRating.description, /Calibrated PR quality rating/);
-  assert.match(prRating.description, /Rate the PR evidence and patch quality/);
-  assert.match(prRating.description, /Do not lower any tier solely because the PR is draft/);
-  assert.match(prRating.description, /has protected labels/);
-  assert.match(prRating.description, /not automerge-eligible/);
-  assert.match(
-    prRating.properties.overallTier.description,
-    /Draft, protected-label, automerge eligibility, and maintainer-waiting workflow states must not lower this tier by themselves/,
-  );
-});
-
-test("review prompt and schema describe positive-only feature showcase labels", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-  const schema = JSON.parse(readFileSync("schema/clawsweeper-decision.schema.json", "utf8"));
-  const featureShowcase = schema.properties.featureShowcase;
-
-  assert.match(prompt, /featureShowcase/);
-  assert.match(prompt, /positive-only maintainer spotlight/);
-  assert.match(prompt, /really compelling feature ideas/);
-  assert.match(prompt, /not a merge gate/);
-  assert.match(featureShowcase.description, /Positive-only maintainer spotlight/);
-  assert.match(featureShowcase.description, /not a merge gate/);
-  assert.deepEqual(featureShowcase.properties.status.enum, ["showcase", "none"]);
-});
-
-test("review prompt classifies Telegram visible proof candidates", () => {
-  const prompt = readFileSync("prompts/review-item.md", "utf8");
-
-  assert.match(prompt, /telegramVisibleProof/);
-  assert.match(prompt, /telegram-crabbox-e2e-proof/);
-  assert.match(prompt, /message formatting/);
-  assert.match(prompt, /mantis: telegram-visible-proof/);
-  assert.match(prompt, /mantisRecommendation/);
-  assert.match(prompt, /@openclaw-mantis/);
-  assert.match(prompt, /ambiguous Mantis account mention/);
-});
-
-test("pull request review comments suggest copy-paste Mantis proof comments", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "83140",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def456",
-    })}
-
-## Summary
-
-Keep this Telegram PR open for maintainer review.
-
-## What This Changes
-
-Fixes Telegram topic stop targeting.
-
-## Real Behavior Proof
-
-Status: mock_only
-
-Evidence kind: none
-
-Needs contributor action: true
-
-Summary: Current proof is test-only for visible Telegram topic behavior.
-
-## Mantis Recommendation
-
-Status: recommended
-
-Scenario: telegram_desktop_proof
-
-Reason: This changes visible Telegram topic behavior that should be proven in native Telegram Desktop.
-
-Maintainer comment: @openclaw-mantis telegram desktop proof: verify that /stop targets the active topic and does not affect other topics.
-
-## Work Candidate
-
-Candidate: none
-
-Confidence: low
-
-Priority: low
-
-Status: none
-
-Reason: Maintainers should review the proof before merge.
-	`,
-    "none",
-  );
-
-  assert.match(comment, /\*\*Mantis proof suggestion\*\*/);
-  assert.match(comment, /posting this exact PR comment/);
-  assert.match(comment, /```text\n@openclaw-mantis telegram desktop proof:/);
-});
-
-test("pull request review comments suppress unsafe Mantis recommendations", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "83140",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def456",
-    })}
-
-## Summary
-
-Keep this Telegram PR open for maintainer review.
-
-## Mantis Recommendation
-
-Status: recommended
-
-Scenario: telegram_desktop_proof
-
-Reason: This changes visible Telegram behavior.
-
-Maintainer comment: @${"mantis"} telegram desktop proof
-
-## Work Candidate
-
-Candidate: none
-
-Confidence: low
-
-Priority: low
-
-Status: none
-
-Reason: Maintainers should review the proof before merge.
-	`,
-    "none",
-  );
-
-  assert.doesNotMatch(comment, /\*\*Mantis proof suggestion\*\*/);
-  assert.doesNotMatch(comment, /@openclaw-mantis/);
-});
-
-test("ClawSweeper proof judgement controls the sufficient proof label", () => {
-  assert.deepEqual(realBehaviorProofSufficientLabelsForTest(["proof: supplied"], "sufficient"), [
-    "proof: supplied",
-    "proof: sufficient",
-  ]);
-  assert.deepEqual(
-    realBehaviorProofSufficientLabelsForTest(
-      ["proof: supplied", "proof: sufficient"],
-      "insufficient",
-    ),
-    ["proof: supplied"],
-  );
-  assert.deepEqual(realBehaviorProofSufficientLabelsForTest(["proof: sufficient"], "missing"), []);
-});
-
-test("ClawSweeper proof evidence kind controls media proof labels", () => {
-  assert.deepEqual(realBehaviorProofMediaLabelsForTest(["bug"], "screenshot"), [
-    "bug",
-    "proof: 📸 screenshot",
-  ]);
-  assert.deepEqual(realBehaviorProofMediaLabelsForTest(["proof: 📸 screenshot"], "recording"), [
-    "proof: 🎥 video",
-  ]);
-  assert.deepEqual(
-    realBehaviorProofMediaLabelsForTest(["proof: 📸 screenshot", "proof: 🎥 video"], "terminal"),
-    [],
-  );
-});
-
-test("ClawSweeper proof label sync recognizes missing optional labels", () => {
-  assert.equal(
-    isMissingGitHubLabelErrorForTest(
-      "failed to update https://github.com/openclaw/fs-safe/pull/18: 'proof: sufficient' not found",
-      "proof: sufficient",
-    ),
-    true,
-  );
-  assert.equal(
-    isMissingGitHubLabelErrorForTest(
-      "failed to update https://github.com/openclaw/fs-safe/pull/18: 'other label' not found",
-      "proof: sufficient",
-    ),
-    false,
-  );
-});
-
-test("ClawSweeper optional label sync recognizes GitHub label capacity errors", () => {
-  assert.equal(
-    isGitHubLabelCapacityErrorForTest(
-      "GraphQL: Validation failed: Labels can have a maximum of 100 labels (addLabelsToLabelable)",
-    ),
-    true,
-  );
-  assert.equal(
-    isGitHubLabelCapacityErrorForTest("GraphQL: Resource not accessible by integration"),
-    false,
-  );
-});
-
-function proofNudgeReport(overrides = {}) {
-  const values = {
-    labels: JSON.stringify(["triage: needs-real-behavior-proof"]),
-    authorAssociation: "CONTRIBUTOR",
-    author: "contributor",
-    reviewStatus: "complete",
-    headSha: "abc123def456",
-    reviewedAt: "2026-01-01T00:00:00Z",
-    itemCategory: "feature",
-    pullFiles: JSON.stringify(["src/app.ts"]),
-    proofStatus: "missing",
-    evidenceKind: "none",
-    needsContributorAction: "true",
-    proofSummary: "The PR needs after-fix proof from a real setup.",
-    securityReviewStatus: "cleared",
-    securityReviewSummary: "No security-sensitive review is needed for this proof nudge.",
-    mantisStatus: "not_recommended",
-    mantisScenario: "none",
-    mantisReason: "Mantis proof is not useful for this sample.",
-    mantisComment: "",
-    ...overrides,
-  };
-  return `---
-repository: openclaw/openclaw
-number: 42
-type: pull_request
-title: Proof nudge sample
-author: ${values.author}
-author_association: ${values.authorAssociation}
-labels: ${values.labels}
-item_category: ${values.itemCategory}
-review_status: ${values.reviewStatus}
-pull_files: ${values.pullFiles}
-pull_files_truncated: false
-pull_head_sha: ${values.headSha}
-reviewed_at: ${values.reviewedAt}
----
-
-## Real Behavior Proof
-
-Status: ${values.proofStatus}
-
-Evidence kind: ${values.evidenceKind}
-
-Needs contributor action: ${values.needsContributorAction}
-
-Summary: ${values.proofSummary}
-
-## Security Review
-
-Status: ${values.securityReviewStatus}
-
-Summary: ${values.securityReviewSummary}
-
-## Mantis Recommendation
-
-Status: ${values.mantisStatus}
-
-Scenario: ${values.mantisScenario}
-
-Reason: ${values.mantisReason}
-
-Maintainer comment: ${values.mantisComment}
-`;
-}
-
-function proofNudgeItem(overrides = {}) {
-  return item({
-    kind: "pull_request",
-    number: 42,
-    title: "Proof nudge sample",
-    author: "contributor",
-    authorAssociation: "CONTRIBUTOR",
-    labels: ["triage: needs-real-behavior-proof"],
-    locked: false,
-    activeLockReason: null,
-    ...overrides,
-  });
-}
-
-test("bot proof candidate scan prioritizes ClawSweeper proof blockers", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    writeFileSync(
-      join(root, "41.md"),
-      proofNudgeReport({
-        number: 41,
-        author: "contributor",
-        reviewedAt: "2026-01-05T00:00:00Z",
-      }),
-    );
-    writeFileSync(
-      join(root, "42.md"),
-      proofNudgeReport({
-        author: "app/clawsweeper",
-        reviewedAt: "2026-01-01T00:00:00Z",
-      }),
-    );
-
-    assert.deepEqual(
-      botProofCandidateRecordsForTest(root).map((candidate) => candidate.number),
-      [42, 41],
-    );
-    assert.deepEqual(
-      botProofCandidateRecordsForTest(root, [41]).map((candidate) => candidate.number),
-      [41],
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("bot proof handling selects maintainer or Mantis proof path for ClawSweeper PRs", () => {
-  const baseOptions = {
-    item: proofNudgeItem({
-      author: "app/clawsweeper",
-      labels: ["triage: needs-real-behavior-proof", "status: 📣 needs proof", "stale"],
-    }),
-    headSha: "abc123def456",
-  };
-  const decision = botProofEligibilityForTest({
-    ...baseOptions,
-    markdown: proofNudgeReport({ author: "app/clawsweeper" }),
-  });
-  assert.equal(decision.eligible, true);
-  assert.equal(decision.action, "bot_proof_decision_planned");
-
-  const mantis = botProofEligibilityForTest({
-    ...baseOptions,
-    markdown: proofNudgeReport({
-      author: "app/clawsweeper",
-      mantisStatus: "recommended",
-      mantisScenario: "telegram_desktop_proof",
-      mantisReason: "Native Telegram Desktop proof would show the visible topic behavior.",
-      mantisComment: "@openclaw-mantis telegram desktop proof: verify the topic behavior",
-    }),
-  });
-  assert.equal(mantis.eligible, true);
-  assert.equal(mantis.action, "bot_proof_mantis_request_planned");
-
-  const manualVisual = botProofEligibilityForTest({
-    ...baseOptions,
-    markdown: proofNudgeReport({
-      author: "app/clawsweeper",
-      mantisStatus: "recommended",
-      mantisScenario: "visual_task",
-      mantisReason: "A visual proof task would show the fixed Control UI behavior.",
-      mantisComment: "@openclaw-mantis capture Control UI proof for this PR",
-    }),
-  });
-  assert.equal(manualVisual.eligible, true);
-  assert.equal(manualVisual.action, "bot_proof_decision_planned");
-
-  assert.equal(
-    botProofEligibilityForTest({
-      item: proofNudgeItem({
-        author: "clawsweeper[bot]",
-        labels: ["status: 📣 needs proof"],
-      }),
-      markdown: proofNudgeReport({ author: "app/clawsweeper" }),
-      headSha: "abc123def456",
-    }).action,
-    "bot_proof_decision_planned",
-  );
-});
-
-test("bot proof handling skips overrides, stale heads, drafts, and non-ClawSweeper PRs", () => {
-  const baseItem = proofNudgeItem({
-    author: "app/clawsweeper",
-    labels: ["triage: needs-real-behavior-proof"],
-  });
-  assert.equal(
-    botProofEligibilityForTest({
-      item: proofNudgeItem({ author: "contributor" }),
-      markdown: proofNudgeReport(),
-      headSha: "abc123def456",
-    }).action,
-    "skipped_not_bot_authored",
-  );
-  assert.equal(
-    botProofEligibilityForTest({
-      item: baseItem,
-      markdown: proofNudgeReport({ author: "app/clawsweeper" }),
-      headSha: "abc123def456",
-      draft: true,
-    }).action,
-    "skipped_draft",
-  );
-  for (const label of ["proof: override", "proof: supplied", "proof: sufficient"] as const) {
-    assert.equal(
-      botProofEligibilityForTest({
-        item: proofNudgeItem({
-          author: "app/clawsweeper",
-          labels: ["triage: needs-real-behavior-proof", label],
-        }),
-        markdown: proofNudgeReport({ author: "app/clawsweeper" }),
-        headSha: "abc123def456",
-      }).action,
-      "skipped_policy_exempt",
-    );
-  }
-  assert.equal(
-    botProofEligibilityForTest({
-      item: baseItem,
-      markdown: proofNudgeReport({ author: "app/clawsweeper", headSha: "oldhead" }),
-      headSha: "newhead",
-    }).action,
-    "skipped_stale_report_head",
-  );
-});
-
-test("bot proof status comment asks maintainers without contributor nudge copy", () => {
-  const markdown = proofNudgeReport({
-    author: "app/clawsweeper",
-    mantisStatus: "recommended",
-    mantisScenario: "visual_task",
-    mantisReason: "A visual proof task would show the fixed Control UI behavior.",
-    mantisComment: "@openclaw-mantis capture Control UI proof for this PR",
-  });
-  const comment = renderBotProofDecisionCommentForTest({
-    number: 42,
-    headSha: "abc123def456",
-    markdown,
-  });
-
-  assert.match(comment, /ClawSweeper-authored replacement PR is blocked on real behavior proof/);
-  assert.match(comment, /proof: override/);
-  assert.match(comment, /Possible manual Mantis\/desktop proof suggestion/);
-  assert.match(comment, /@openclaw-mantis capture Control UI proof/);
-  assert.match(comment, /<!-- clawsweeper-bot-proof-decision item="42" sha="abc123def456"/);
-  assert.doesNotMatch(comment, /thanks for the PR/);
-  assert.doesNotMatch(comment, /Once proof is added/);
-});
-
-test("proof nudge candidate scan defers proof-label checks to live PR state", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    writeFileSync(join(root, "41.md"), proofNudgeReport({ reviewedAt: "2026-01-05T00:00:00Z" }));
-    writeFileSync(
-      join(root, "42.md"),
-      `---
-repository: openclaw/openclaw
-number: 42
-type: pull_request
-labels: []
-review_status: complete
-reviewed_at: 2026-01-01T00:00:00Z
----
-
-## Summary
-
-Stored report label snapshots can be older than the live PR labels.
-`,
-    );
-
-    const requestedCandidates = proofNudgeCandidateRecordsForTest(root, [42]);
-    const allCandidates = proofNudgeCandidateRecordsForTest(root);
-
-    assert.deepEqual(
-      requestedCandidates.map((candidate) => candidate.number),
-      [42],
-    );
-    assert.deepEqual(
-      allCandidates.map((candidate) => candidate.number),
-      [41, 42],
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("proof nudges skip stale targeted PRs when live GitHub state cannot be fetched", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const reportPath = join(root, "proof-nudge-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    writeFileSync(join(itemsDir, "42.md"), proofNudgeReport({ number: 42 }));
-    writeFileSync(join(itemsDir, "43.md"), proofNudgeReport({ number: 43 }));
-    withMockGh(
-      root,
-      `#!/usr/bin/env node
-const args = process.argv.slice(2);
-const path = args[1] || "";
-if (path.endsWith("/issues/42")) {
-  console.error("Not Found");
-  process.exit(1);
-}
-if (path.endsWith("/issues/43")) {
-  console.log(JSON.stringify({
-    number: 43,
-    title: "Closed proof nudge sample",
-    html_url: "https://github.com/openclaw/openclaw/pull/43",
-    created_at: "2026-01-01T00:00:00Z",
-    updated_at: "2026-01-02T00:00:00Z",
-    closed_at: "2026-01-03T00:00:00Z",
-    state: "closed",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "contributor" },
-    labels: ["triage: needs-real-behavior-proof"],
-    pull_request: {}
-  }));
-  process.exit(0);
-}
-console.error("unexpected gh args: " + args.join(" "));
-process.exit(1);
-`,
-      () => {
-        execFileSync(process.execPath, [
-          "dist/clawsweeper.js",
-          "proof-nudges",
-          "--target-repo",
-          "openclaw/openclaw",
-          "--items-dir",
-          itemsDir,
-          "--item-numbers",
-          "42,43",
-          "--limit",
-          "10",
-          "--processed-limit",
-          "10",
-          "--report-path",
-          reportPath,
-        ]);
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8"));
-    assert.deepEqual(
-      report.map((entry: { number: number; action: string }) => [entry.number, entry.action]),
-      [
-        [42, "skipped_live_fetch_failed"],
-        [43, "skipped_not_open"],
-      ],
-    );
-    assert.match(report[0].reason, /live GitHub state could not be fetched/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("proof nudges become eligible only after proof policy and age gates pass", () => {
-  const result = proofNudgeEligibilityForTest({
-    item: proofNudgeItem(),
-    markdown: proofNudgeReport(),
-    headSha: "abc123def456",
-    headCommittedAt: "2026-01-01T00:00:00Z",
-    now: Date.parse("2026-01-10T00:00:00Z"),
-    minAgeDays: 5,
-    cooldownDays: 7,
-  });
-
-  assert.equal(result.eligible, true);
-  assert.equal(result.action, "proof_nudge_planned");
-});
-
-test("proof nudges skip recent reviews and recent author activity", () => {
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem(),
-      markdown: proofNudgeReport({ reviewedAt: "2026-01-08T00:00:00Z" }),
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-      minAgeDays: 5,
-      cooldownDays: 7,
-    }).action,
-    "skipped_recent_review",
-  );
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem(),
-      markdown: proofNudgeReport(),
-      comments: [
-        {
-          author: "contributor",
-          body: "I'll add proof later.",
-          updatedAt: "2026-01-09T00:00:00Z",
-        },
-      ],
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-      minAgeDays: 5,
-      cooldownDays: 7,
-    }).action,
-    "skipped_recent_author_activity",
-  );
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem(),
-      markdown: proofNudgeReport(),
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      authorEditedAt: "2026-01-09T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-      minAgeDays: 5,
-      cooldownDays: 7,
-    }).action,
-    "skipped_recent_author_activity",
-  );
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem(),
-      markdown: proofNudgeReport(),
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      authorReviewActivityAt: "2026-01-09T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-      minAgeDays: 5,
-      cooldownDays: 7,
-    }).action,
-    "skipped_recent_author_activity",
-  );
-  const maintainerActivity = proofNudgeEligibilityForTest({
-    item: proofNudgeItem({ updatedAt: "2026-01-09T00:00:00Z" }),
-    markdown: proofNudgeReport(),
-    comments: [
-      { author: "maintainer", body: "Queued for review.", updatedAt: "2026-01-09T00:00:00Z" },
-    ],
-    headSha: "abc123def456",
-    headCommittedAt: "2026-01-01T00:00:00Z",
-    now: Date.parse("2026-01-10T00:00:00Z"),
-    minAgeDays: 5,
-    cooldownDays: 7,
-  });
-
-  assert.equal(maintainerActivity.eligible, true);
-  assert.equal(maintainerActivity.action, "proof_nudge_planned");
-});
-
-test("proof nudges use same-head cooldown markers instead of label churn", () => {
-  const comment = renderProofNudgeCommentForTest({
-    number: 42,
-    author: "contributor",
-    headSha: "abc123def456",
-    timestamp: "2026-01-08T00:00:00.000Z",
-  });
-  const result = proofNudgeEligibilityForTest({
-    item: proofNudgeItem(),
-    markdown: proofNudgeReport(),
-    comments: [{ author: "clawsweeper[bot]", body: comment, updatedAt: "2026-01-08T00:00:00Z" }],
-    headSha: "abc123def456",
-    headCommittedAt: "2026-01-01T00:00:00Z",
-    now: Date.parse("2026-01-10T00:00:00Z"),
-    minAgeDays: 5,
-    cooldownDays: 7,
-  });
-
-  assert.equal(result.eligible, false);
-  assert.equal(result.action, "skipped_recent_nudge");
-  assert.match(comment, /<!-- clawsweeper-proof-nudge item="42" sha="abc123def456"/);
-
-  const spoofedMarker = proofNudgeEligibilityForTest({
-    item: proofNudgeItem(),
-    markdown: proofNudgeReport(),
-    comments: [{ author: "someoneelse", body: comment, updatedAt: "2026-01-08T00:00:00Z" }],
-    headSha: "abc123def456",
-    headCommittedAt: "2026-01-01T00:00:00Z",
-    now: Date.parse("2026-01-10T00:00:00Z"),
-    minAgeDays: 5,
-    cooldownDays: 7,
-  });
-
-  assert.equal(spoofedMarker.eligible, true);
-  assert.equal(spoofedMarker.action, "proof_nudge_planned");
-
-  const malformedMarker = proofNudgeEligibilityForTest({
-    item: proofNudgeItem(),
-    markdown: proofNudgeReport(),
-    comments: [
-      {
-        author: "clawsweeper[bot]",
-        body: `<!-- clawsweeper-proof-nudge ${"a".repeat(5000)} -->`,
-        updatedAt: "2026-01-08T00:00:00Z",
-      },
-    ],
-    headSha: "abc123def456",
-    headCommittedAt: "2026-01-01T00:00:00Z",
-    now: Date.parse("2026-01-10T00:00:00Z"),
-    minAgeDays: 5,
-    cooldownDays: 7,
-  });
-
-  assert.equal(malformedMarker.eligible, true);
-  assert.equal(malformedMarker.action, "proof_nudge_planned");
-});
-
-test("proof nudges skip stale report heads and proof-policy exemptions", () => {
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem(),
-      markdown: proofNudgeReport({ headSha: "oldhead" }),
-      headSha: "newhead",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-    }).action,
-    "skipped_stale_report_head",
-  );
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem(),
-      markdown: proofNudgeReport({ headSha: "unknown" }),
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-    }).action,
-    "skipped_stale_report_head",
-  );
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem({ labels: ["triage: needs-real-behavior-proof", "proof: override"] }),
-      markdown: proofNudgeReport({
-        labels: JSON.stringify(["triage: needs-real-behavior-proof", "proof: override"]),
-      }),
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-    }).action,
-    "skipped_policy_exempt",
-  );
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem(),
-      markdown: proofNudgeReport({
-        proofStatus: "not_applicable",
-        needsContributorAction: "false",
-        reviewStatus: "failed",
-      }),
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-    }).action,
-    "skipped_policy_exempt",
-  );
-  assert.equal(
-    proofNudgeEligibilityForTest({
-      item: proofNudgeItem({ labels: ["triage: needs-real-behavior-proof", "proof: supplied"] }),
-      markdown: proofNudgeReport(),
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-    }).action,
-    "skipped_policy_exempt",
-  );
-});
-
-test("proof nudges skip maintainer, bot, security, and release PRs", () => {
-  for (const [input, expected] of [
-    [proofNudgeItem({ authorAssociation: "MEMBER" }), "skipped_maintainer_authored"],
-    [proofNudgeItem({ author: "dependabot[bot]" }), "skipped_bot_authored"],
-    [
-      proofNudgeItem({
-        labels: ["triage: needs-real-behavior-proof", "clawsweeper:needs-security-review"],
-      }),
-      "skipped_protected_label",
-    ],
-    [proofNudgeItem({ title: "Release 1.2.3" }), "skipped_protected_label"],
-    [proofNudgeItem({ title: "chore(release): 1.2.3" }), "skipped_protected_label"],
-  ] as const) {
-    assert.equal(
-      proofNudgeEligibilityForTest({
-        item: input,
-        markdown: proofNudgeReport({
-          author: input.author,
-          authorAssociation: input.authorAssociation,
-          labels: JSON.stringify(input.labels),
-        }),
-        headSha: "abc123def456",
-        headCommittedAt: "2026-01-01T00:00:00Z",
-        now: Date.parse("2026-01-10T00:00:00Z"),
-      }).action,
-      expected,
-    );
-  }
-
-  for (const [markdown, expectedReason] of [
-    [
-      proofNudgeReport({
-        itemCategory: "security",
-      }),
-      "latest report item category is security",
-    ],
-    [
-      proofNudgeReport({
-        securityReviewStatus: "needs_attention",
-        securityReviewSummary: "The patch changes token handling and needs maintainer review.",
-      }),
-      "latest report security review needs attention",
-    ],
-  ] as const) {
-    const result = proofNudgeEligibilityForTest({
-      item: proofNudgeItem(),
-      markdown,
-      headSha: "abc123def456",
-      headCommittedAt: "2026-01-01T00:00:00Z",
-      now: Date.parse("2026-01-10T00:00:00Z"),
-    });
-    assert.equal(result.action, "skipped_protected_label");
-    assert.equal(result.reason, expectedReason);
-  }
-});
-
-test("proof nudges skip locked PR conversations before posting", () => {
-  const result = proofNudgeEligibilityForTest({
-    item: proofNudgeItem({ locked: true, activeLockReason: "resolved" }),
-    markdown: proofNudgeReport(),
-    headSha: "abc123def456",
-    headCommittedAt: "2026-01-01T00:00:00Z",
-    now: Date.parse("2026-01-10T00:00:00Z"),
-  });
-
-  assert.equal(result.eligible, false);
-  assert.equal(result.action, "skipped_locked_conversation");
-  assert.match(result.reason, /conversation is locked \(resolved\)/);
-});
-
-test("proof nudge copy asks for evidence without promising author-only re-review access", () => {
-  const comment = renderProofNudgeCommentForTest({
-    number: 42,
-    author: "contributor",
-    headSha: "abc123def456",
-  });
-
-  assert.match(comment, /^@contributor thanks for the PR\./);
-  assert.match(comment, /screenshot, short video, terminal output/);
-  assert.match(comment, /ClawSweeper or a maintainer can re-check it/);
-  assert.doesNotMatch(comment, /@clawsweeper re-review/);
-});
-
-test("ClawSweeper PR rating labels use one themed overall label", () => {
-  assert.deepEqual(prRatingLabelsForTest(["bug"], "A"), ["bug", "rating: 🦞 diamond lobster"]);
-  assert.deepEqual(
-    prRatingLabelsForTest(["rating: 🦀 challenger crab", "bug", "rating: 🦐 gold shrimp"], "D"),
-    ["bug", "rating: 🦪 silver shellfish"],
-  );
-  assert.deepEqual(prRatingLabelsForTest(["bug"], "bogus"), [
-    "bug",
-    "rating: 🌊 off-meta tidepool",
-  ]);
-  assert.deepEqual(prRatingLabelsForTest(["bug", "rating: 🌊 off-meta tidepool"], "NA", true), [
-    "bug",
-  ]);
-});
-
-test("ClawSweeper PR rating label scheme exposes boring internal tiers", () => {
-  assert.deepEqual(
-    prRatingLabelSchemeForTest().map(({ tier, name, color }) => ({ tier, name, color })),
-    [
-      { tier: "S", name: "rating: 🦀 challenger crab", color: "1F883D" },
-      { tier: "A", name: "rating: 🦞 diamond lobster", color: "0969DA" },
-      { tier: "B", name: "rating: 🐚 platinum hermit", color: "0F766E" },
-      { tier: "C", name: "rating: 🦐 gold shrimp", color: "B7791F" },
-      { tier: "D", name: "rating: 🦪 silver shellfish", color: "7A828E" },
-      { tier: "F", name: "rating: 🧂 unranked krab", color: "8C2F39" },
-      { tier: "NA", name: "rating: 🌊 off-meta tidepool", color: "6E7781" },
-    ],
-  );
-});
-
-test("ClawSweeper feature showcase label is positive-only and high signal", () => {
-  assert.deepEqual(
-    featureShowcaseLabelsForTest(["enhancement"], {
-      itemCategory: "feature",
-      status: "showcase",
-      securityReviewStatus: "cleared",
-      overallCorrectness: "patch is correct",
-    }),
-    ["enhancement", "feature: ✨ showcase"],
-  );
-  assert.deepEqual(
-    featureShowcaseLabelsForTest(["enhancement"], {
-      itemCategory: "feature",
-      status: "none",
-      securityReviewStatus: "cleared",
-      overallCorrectness: "patch is correct",
-    }),
-    ["enhancement"],
-  );
-  assert.deepEqual(
-    featureShowcaseLabelsForTest(["feature: ✨ showcase"], {
-      itemCategory: "feature",
-      status: "none",
-      securityReviewStatus: "cleared",
-      overallCorrectness: "patch is correct",
-    }),
-    ["feature: ✨ showcase"],
-  );
-});
-
-test("ClawSweeper feature showcase label does not apply to unsafe or non-feature PRs", () => {
-  assert.deepEqual(
-    featureShowcaseLabelsForTest(["bug"], {
-      itemCategory: "bug",
-      status: "showcase",
-      securityReviewStatus: "cleared",
-      overallCorrectness: "patch is correct",
-    }),
-    ["bug"],
-  );
-  assert.deepEqual(
-    featureShowcaseLabelsForTest(["enhancement"], {
-      itemCategory: "feature",
-      status: "showcase",
-      securityReviewStatus: "needs_attention",
-      overallCorrectness: "patch is correct",
-    }),
-    ["enhancement"],
-  );
-  assert.deepEqual(
-    featureShowcaseLabelsForTest(["enhancement"], {
-      itemCategory: "feature",
-      status: "showcase",
-      securityReviewStatus: "cleared",
-      overallCorrectness: "patch is incorrect",
-    }),
-    ["enhancement"],
-  );
-});
-
-test("ClawSweeper PR status labels use one current workflow status", () => {
-  assert.deepEqual(
-    prStatusLabelsForTest(["bug", "status: ⏳ waiting on author"], {
-      findingPriorities: [2],
-      hasRecentAuthorActivity: true,
-    }),
-    ["bug", "status: 🛠️ actively grinding"],
-  );
-  assert.deepEqual(
-    prStatusLabelsForTest(["bug", "status: 🛠️ actively grinding"], {
-      proofStatus: "sufficient",
-      overallCorrectness: "patch is correct",
-    }),
-    ["bug", "status: 👀 ready for maintainer look"],
-  );
-});
-
-test("ClawSweeper PR status routes security owner acceptance to maintainer look", () => {
-  const ownerAcceptanceLabels = prStatusLabelsForTest([], {
-    proofStatus: "sufficient",
-    securityStatus: "needs_attention",
-    mergeRiskOptions: [{ category: "accept_risk", recommended: true }],
-    overallCorrectness: "patch is correct",
-  });
-
-  assert.equal(
-    ownerAcceptanceLabels.some((label) => label.endsWith("ready for maintainer look")),
-    true,
-  );
-  assert.equal(
-    ownerAcceptanceLabels.some((label) => label.endsWith("waiting on author")),
-    false,
-  );
-
-  const authorFixLabels = prStatusLabelsForTest([], {
-    proofStatus: "sufficient",
-    securityStatus: "needs_attention",
-    mergeRiskOptions: [{ category: "fix_before_merge", recommended: true }],
-    overallCorrectness: "patch is correct",
-  });
-
-  assert.equal(
-    authorFixLabels.some((label) => label.endsWith("waiting on author")),
-    true,
-  );
-
-  const ambiguousSecurityLabels = prStatusLabelsForTest([], {
-    proofStatus: "sufficient",
-    securityStatus: "needs_attention",
-    overallCorrectness: "patch is correct",
-  });
-
-  assert.equal(
-    ambiguousSecurityLabels.some((label) => label.endsWith("waiting on author")),
-    true,
-  );
-});
-
-test("ClawSweeper PR status labels preserve other label families", () => {
-  assert.deepEqual(
-    prStatusLabelsForTest(
-      [
-        "rating: 🦞 diamond lobster",
-        "merge-risk: 🚨 compatibility",
-        "proof: sufficient",
-        "status: custom-user-label",
-      ],
-      {
-        proofStatus: "missing",
-      },
-    ),
-    [
-      "rating: 🦞 diamond lobster",
-      "merge-risk: 🚨 compatibility",
-      "proof: sufficient",
-      "status: custom-user-label",
-      "status: 📣 needs proof",
-    ],
-  );
-});
-
-test("ClawSweeper PR status labels respect priority ordering", () => {
-  assert.deepEqual(
-    prStatusLabelsForTest(["clawsweeper:automerge"], {
-      proofStatus: "missing",
-      hasRecentReReviewRequest: true,
-    }),
-    ["clawsweeper:automerge", "status: 🚀 automerge armed"],
-  );
-  assert.deepEqual(
-    prStatusLabelsForTest([], {
-      proofStatus: "missing",
-      hasRecentAuthorActivity: true,
-      hasRecentReReviewRequest: true,
-    }),
-    ["status: 🔁 re-review loop"],
-  );
-  assert.deepEqual(
-    prStatusLabelsForTest([], {
-      proofStatus: "missing",
-      hasRecentAuthorActivity: true,
-    }),
-    ["status: 🛠️ actively grinding"],
-  );
-  assert.deepEqual(
-    prStatusLabelsForTest([], {
-      proofStatus: "missing",
-    }),
-    ["status: 📣 needs proof"],
-  );
-  assert.deepEqual(
-    prStatusLabelsForTest([], {
-      findingPriorities: [2],
-    }),
-    ["status: ⏳ waiting on author"],
-  );
-});
-
-test("ClawSweeper PR status ignores bot-authored re-review guidance", () => {
-  assert.deepEqual(
-    prStatusLabelsForTest([], {
-      proofStatus: "missing",
-      reviewedAt: "2026-01-01T00:00:00Z",
-      comments: [
-        {
-          author: "openclaw-clawsweeper[bot]",
-          body: "After adding proof, comment `@clawsweeper re-review`.",
-          updatedAt: "2026-01-01T00:01:00Z",
-        },
-      ],
-    }),
-    ["status: 📣 needs proof"],
-  );
-  assert.deepEqual(
-    prStatusLabelsForTest([], {
-      proofStatus: "missing",
-      reviewedAt: "2026-01-01T00:00:00Z",
-      comments: [
-        {
-          author: "contributor",
-          body: "@clawsweeper re-review",
-          createdAt: "2026-01-01T00:01:00Z",
-        },
-      ],
-    }),
-    ["status: 🔁 re-review loop"],
-  );
-});
-
-test("ClawSweeper PR status treats maintainer-only rank-up moves as ready", () => {
-  assert.deepEqual(
-    prStatusLabelsForTest([], {
-      nextSteps: [
-        "Maintainer accepts the relative details.reportPath contract change before merge.",
-      ],
-      proofStatus: "sufficient",
-      overallCorrectness: "patch is correct",
-    }),
-    ["status: 👀 ready for maintainer look"],
-  );
-});
-
-test("ClawSweeper PR status labels are PR-only", () => {
-  assert.deepEqual(
-    prStatusLabelsForTest(["bug", "status: ⏳ waiting on author"], {
-      isPullRequest: false,
-      nextSteps: ["Add proof."],
-    }),
-    ["bug"],
-  );
-});
-
-test("ClawSweeper PR status label scheme exposes workflow states", () => {
-  assert.deepEqual(
-    prStatusLabelSchemeForTest().map(({ kind, name, color }) => ({ kind, name, color })),
-    [
-      { kind: "automerge_armed", name: "status: 🚀 automerge armed", color: "0E8A16" },
-      { kind: "re_review_loop", name: "status: 🔁 re-review loop", color: "8250DF" },
-      { kind: "actively_grinding", name: "status: 🛠️ actively grinding", color: "0969DA" },
-      { kind: "needs_proof", name: "status: 📣 needs proof", color: "D93F0B" },
-      {
-        kind: "needs_maintainer_proof_decision",
-        name: "status: needs maintainer proof decision",
-        color: "D93F0B",
-      },
-      { kind: "waiting_on_author", name: "status: ⏳ waiting on author", color: "FBCA04" },
-      {
-        kind: "ready_for_maintainer_look",
-        name: "status: 👀 ready for maintainer look",
-        color: "2DA44E",
-      },
-    ],
-  );
-});
-
-test("ClawSweeper Telegram proof judgement controls the Mantis proof label", () => {
-  assert.deepEqual(telegramVisibleProofLabelsForTest(["channel: telegram"], "needed"), [
-    "channel: telegram",
-    "mantis: telegram-visible-proof",
-  ]);
-  assert.deepEqual(
-    telegramVisibleProofLabelsForTest(
-      ["channel: telegram", "mantis: telegram-visible-proof"],
-      "not_needed",
-    ),
-    ["channel: telegram"],
-  );
-});
-
-test("ClawSweeper priority label scheme exposes P0 through P3 labels", () => {
-  assert.deepEqual(priorityLabelSchemeForTest(), [
-    {
-      name: "P0",
-      color: "B60205",
-      description: "Emergency: data loss, security bypass, crash loop, or unusable core runtime.",
-    },
-    {
-      name: "P1",
-      color: "D93F0B",
-      description: "Urgent regression or broken agent/channel workflow affecting real users now.",
-    },
-    {
-      name: "P2",
-      color: "FBCA04",
-      description: "Normal priority bug or improvement with limited blast radius.",
-    },
-    {
-      name: "P3",
-      color: "8C959F",
-      description: "Low-risk cleanup, docs, polish, ergonomics, or speculative feature.",
-    },
-  ]);
-});
-
-test("ClawSweeper priority label descriptions fit GitHub label limits", () => {
-  for (const label of priorityLabelSchemeForTest()) {
-    assert.ok(
-      label.description.length <= 100,
-      `${label.name} description is ${label.description.length} characters`,
-    );
-  }
-});
-
-test("ClawSweeper priority label descriptions stay aligned with prompt and schema", () => {
-  const schema = JSON.parse(reviewDecisionSchemaText()) as {
-    properties?: {
-      triagePriority?: {
-        description?: string;
-      };
-    };
-  };
-  const schemaDescription = schema.properties?.triagePriority?.description ?? "";
-  const prompt = reviewPromptTemplate();
-  for (const label of priorityLabelSchemeForTest()) {
-    assert.ok(
-      prompt.includes(`\`${label.name}\`: ${label.description}`),
-      `${label.name} description is missing from the review prompt`,
-    );
-    assert.ok(
-      schemaDescription.includes(`${label.name}: ${label.description}`),
-      `${label.name} description is missing from the schema`,
-    );
-  }
-});
-
-test("review prompt keeps unrelated CI noise out of triage priority", () => {
-  const schema = JSON.parse(reviewDecisionSchemaText()) as {
-    properties?: {
-      triagePriority?: {
-        description?: string;
-      };
-    };
-  };
-  const schemaDescription = schema.properties?.triagePriority?.description ?? "";
-  const prompt = reviewPromptTemplate();
-
-  assert.match(prompt, /Do not raise `triagePriority` solely because CI or status checks/);
-  assert.match(
-    prompt,
-    /failing,\s+pending,\s+missing,\s+flaky,\s+or require routine maintainer follow-up/,
-  );
-  assert.match(prompt, /PR diff plausibly caused an urgent regression/);
-  assert.match(schemaDescription, /Do not raise priority solely because CI or status checks/);
-  assert.match(schemaDescription, /diff-caused urgent regressions/);
-});
-
-test("ClawSweeper priority labels follow triage priority", () => {
-  assert.deepEqual(priorityLabelsForTest(["bug"], "P2"), ["bug", "P2"]);
-  assert.deepEqual(priorityLabelsForTest(["bug", "P3"], "P1"), ["bug", "P1"]);
-  assert.deepEqual(priorityLabelsForTest(["P0", "bug"], "none"), ["bug"]);
-});
-
-test("ClawSweeper label justifications render selected label reasons", () => {
-  assert.equal(
-    labelJustificationsMarkdownForTest([
-      {
-        label: "P1",
-        reason: "The PR changes an active channel workflow affecting real users.",
-      },
-      {
-        label: "impact:message-loss",
-        reason: "The diff touches message retry and delivery ordering.",
-      },
-      {
-        label: "merge-risk: 🚨 compatibility",
-        reason: "Merging changes the default upgrade behavior for existing configs.",
-      },
-    ]),
-    [
-      "- `P1`: The PR changes an active channel workflow affecting real users.",
-      "- `impact:message-loss`: The diff touches message retry and delivery ordering.",
-      "- `merge-risk: 🚨 compatibility`: Merging changes the default upgrade behavior for existing configs.",
-    ].join("\n"),
-  );
-});
-
-test("ClawSweeper impact label scheme exposes owned impact labels", () => {
-  assert.deepEqual(impactLabelSchemeForTest(), [
-    {
-      name: "impact:data-loss",
-      color: "B60205",
-      description:
-        "This issue is about lost, corrupted, or silently dropped user/session/config data.",
-    },
-    {
-      name: "impact:security",
-      color: "B60205",
-      description:
-        "This issue is about security boundaries, credentials, authz, sandboxing, or sensitive data.",
-    },
-    {
-      name: "impact:crash-loop",
-      color: "D93F0B",
-      description:
-        "This issue is about crashes, hangs, restart loops, or process-level availability.",
-    },
-    {
-      name: "impact:message-loss",
-      color: "D93F0B",
-      description:
-        "This issue is about lost, duplicated, misrouted, or suppressed channel messages.",
-    },
-    {
-      name: "impact:session-state",
-      color: "F9D65C",
-      description:
-        "This issue is about session, memory, transcript, context, or agent state drift.",
-    },
-    {
-      name: "impact:auth-provider",
-      color: "F9D65C",
-      description:
-        "This issue is about auth, provider routing, model choice, or SecretRef resolution.",
-    },
-    {
-      name: "impact:other",
-      color: "C5DEF5",
-      description:
-        "This issue has meaningful maintainer-visible impact outside the owned taxonomy.",
-    },
-  ]);
-});
-
-test("ClawSweeper impact label descriptions fit GitHub label limits", () => {
-  for (const label of impactLabelSchemeForTest()) {
-    assert.ok(
-      label.description.length <= 100,
-      `${label.name} description is ${label.description.length} characters`,
-    );
-  }
-});
-
-test("ClawSweeper impact label descriptions stay aligned with prompt and schema", () => {
-  const schema = JSON.parse(reviewDecisionSchemaText()) as {
-    properties?: {
-      impactLabels?: {
-        description?: string;
-      };
-    };
-  };
-  const schemaDescription = schema.properties?.impactLabels?.description ?? "";
-  const prompt = reviewPromptTemplate();
-  for (const label of impactLabelSchemeForTest()) {
-    assert.ok(
-      prompt.includes(`\`${label.name}\`: ${label.description}`),
-      `${label.name} description is missing from the review prompt`,
-    );
-    assert.ok(
-      schemaDescription.includes(`${label.name}: ${label.description}`),
-      `${label.name} description is missing from the schema`,
-    );
-  }
-});
-
-test("ClawSweeper impact label schema avoids unsupported response-format keywords", () => {
-  const schema = JSON.parse(reviewDecisionSchemaText()) as {
-    properties?: {
-      impactLabels?: Record<string, unknown>;
-    };
-  };
-  assert.equal(schema.properties?.impactLabels?.uniqueItems, undefined);
-});
-
-test("ClawSweeper merge-risk label scheme exposes PR-only merge warning labels", () => {
-  assert.deepEqual(mergeRiskLabelSchemeForTest(), [
-    {
-      name: "merge-risk: 🚨 compatibility",
-      color: "D1242F",
-      description:
-        "🚨 Merging this PR could break existing users, config, migrations, defaults, or upgrades.",
-    },
-    {
-      name: "merge-risk: 🚨 message-delivery",
-      color: "D1242F",
-      description:
-        "🚨 Merging this PR could drop, duplicate, misroute, suppress, or wrongly target messages.",
-    },
-    {
-      name: "merge-risk: 🚨 session-state",
-      color: "F97316",
-      description:
-        "🚨 Merging this PR could lose, corrupt, stale, or mis-associate session or agent state.",
-    },
-    {
-      name: "merge-risk: 🚨 auth-provider",
-      color: "F97316",
-      description:
-        "🚨 Merging this PR could break OAuth, tokens, provider routing, model choice, or credentials.",
-    },
-    {
-      name: "merge-risk: 🚨 security-boundary",
-      color: "B60205",
-      description:
-        "🚨 Merging this PR could weaken sandboxing, authorization, credentials, or sensitive data.",
-    },
-    {
-      name: "merge-risk: 🚨 availability",
-      color: "D93F0B",
-      description:
-        "🚨 Merging this PR could cause crashes, hangs, restart loops, stalls, or process outages.",
-    },
-    {
-      name: "merge-risk: 🚨 automation",
-      color: "FBCA04",
-      description:
-        "🚨 Merging this PR could break CI, automerge, proof capture, label sync, or automation.",
-    },
-    {
-      name: "merge-risk: 🚨 other",
-      color: "C5DEF5",
-      description: "🚨 Merging this PR has meaningful risk outside the owned taxonomy.",
-    },
-  ]);
-});
-
-test("ClawSweeper merge-risk label descriptions fit GitHub label limits", () => {
-  for (const label of mergeRiskLabelSchemeForTest()) {
-    assert.ok(
-      label.description.length <= 100,
-      `${label.name} description is ${label.description.length} characters`,
-    );
-  }
-});
-
-test("ClawSweeper merge-risk label descriptions stay aligned with prompt and schema", () => {
-  const schema = JSON.parse(reviewDecisionSchemaText()) as {
-    properties?: {
-      mergeRiskLabels?: {
-        description?: string;
-      };
-    };
-  };
-  const schemaDescription = schema.properties?.mergeRiskLabels?.description ?? "";
-  const prompt = reviewPromptTemplate();
-  for (const label of mergeRiskLabelSchemeForTest()) {
-    assert.ok(
-      prompt.includes(`\`${label.name}\`: ${label.description}`),
-      `${label.name} description is missing from the review prompt`,
-    );
-    assert.ok(
-      schemaDescription.includes(`${label.name}: ${label.description}`),
-      `${label.name} description is missing from the schema`,
-    );
-  }
-});
-
-test("review prompt uses automation merge risk only for diff-caused automation risk", () => {
-  const schema = JSON.parse(reviewDecisionSchemaText()) as {
-    properties?: {
-      mergeRiskLabels?: {
-        description?: string;
-      };
-    };
-  };
-  const schemaDescription = schema.properties?.mergeRiskLabels?.description ?? "";
-  const prompt = reviewPromptTemplate();
-
-  assert.match(prompt, /Do not use `merge-risk: 🚨 automation` only because CI is red/);
-  assert.match(prompt, /pending,\s+flaky,\s+or absent/);
-  assert.match(prompt, /PR diff changes automation behavior/);
-  assert.match(prompt, /plausibly causes CI,\s+automerge,\s+proof capture,\s+label sync/);
-  assert.match(schemaDescription, /Do not use merge-risk: 🚨 automation only because CI is red/);
-  assert.match(schemaDescription, /PR diff changes automation behavior/);
-});
-
-test("ClawSweeper merge-risk labels remove stale owned labels and preserve unrelated labels", () => {
-  assert.deepEqual(
-    mergeRiskLabelsForTest(
-      ["bug", "merge-risk: 🚨 compatibility", "merge-risk: 🚨 availability", "impact:message-loss"],
-      ["merge-risk: 🚨 message-delivery", "merge-risk: 🚨 other", "not-a-merge-risk-label"],
-    ),
-    ["bug", "impact:message-loss", "merge-risk: 🚨 message-delivery", "merge-risk: 🚨 other"],
-  );
-  assert.deepEqual(mergeRiskLabelsForTest(["bug", "merge-risk: 🚨 auth-provider"], []), ["bug"]);
-});
-
-test("ClawSweeper impact labels remove stale owned labels and preserve unrelated labels", () => {
-  assert.deepEqual(
-    impactLabelsForTest(
-      ["bug", "impact:data-loss", "impact:security", "proof: sufficient", "P1"],
-      ["impact:message-loss", "impact:other", "not-an-impact-label"],
-    ),
-    ["bug", "proof: sufficient", "P1", "impact:message-loss", "impact:other"],
-  );
-  assert.deepEqual(impactLabelsForTest(["bug", "impact:auth-provider"], []), ["bug"]);
-});
-
-test("ClawSweeper impact labels do not alter PR review finding priorities", () => {
-  const decision = parseDecision(
-    closeDecision({
-      impactLabels: ["impact:data-loss", "impact:security"],
-      labelJustifications: [
-        {
-          label: "P2",
-          reason: "Normal priority applies to this limited-scope implemented behavior check.",
-        },
-        {
-          label: "impact:data-loss",
-          reason: "The selected labels include a data-loss impact classification.",
-        },
-        {
-          label: "impact:security",
-          reason: "The selected labels include a security impact classification.",
-        },
-      ],
-      reviewFindings: [
-        {
-          title: "A concrete review finding",
-          body: "This remains a PR review finding priority, not an impact label.",
-          priority: 1,
-          confidenceScore: 0.9,
-          file: "src/example.ts",
-          lineStart: 10,
-          lineEnd: 10,
-        },
-      ],
-    }),
-  );
-  assert.deepEqual(decision.impactLabels, ["impact:data-loss", "impact:security"]);
-  assert.equal(decision.reviewFindings[0]?.priority, 1);
-});
-
-test("ClawSweeper issue advisory labels expose high-confidence reproduction state", () => {
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      reproductionStatus: "reproduced",
-      reproductionConfidence: "high",
-    }),
-    ["bug", "issue-rating: 🦀 challenger crab", "clawsweeper:current-main-repro"],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      reproductionStatus: "source_reproducible",
-      reproductionConfidence: "high",
-    }),
-    ["bug", "issue-rating: 🦞 diamond lobster", "clawsweeper:source-repro"],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      reproductionStatus: "reproduced",
-      reproductionConfidence: "medium",
-    }),
-    ["bug", "issue-rating: 🐚 platinum hermit"],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      reproductionStatus: "not_reproduced",
-      reproductionConfidence: "high",
-    }),
-    ["bug", "issue-rating: 🦪 silver shellfish", "clawsweeper:not-repro-on-main"],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      reproductionStatus: "source_reproducible",
-      reproductionConfidence: "medium",
-    }),
-    ["bug", "issue-rating: 🐚 platinum hermit", "clawsweeper:needs-live-repro"],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      reproductionStatus: "unclear",
-      reproductionConfidence: "low",
-    }),
-    ["bug", "issue-rating: 🦪 silver shellfish", "clawsweeper:needs-info"],
-  );
-});
-
-test("ClawSweeper issue advisory labels expose work-lane routing state", () => {
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["clawsweeper"], {
-      type: "issue",
-      workCandidate: "queue_fix_pr",
-      workStatus: "candidate",
-      workConfidence: "high",
-      hasWorkShape: true,
-    }),
-    [
-      "clawsweeper",
-      "no-stale",
-      "issue-rating: 🧂 unranked krab",
-      "clawsweeper:queueable-fix",
-      "clawsweeper:fix-shape-clear",
-    ],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["clawsweeper"], {
-      type: "issue",
-      workCandidate: "queue_fix_pr",
-      workStatus: "candidate",
-      workConfidence: "medium",
-    }),
-    ["clawsweeper", "issue-rating: 🧂 unranked krab"],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["clawsweeper"], {
-      type: "issue",
-      workCandidate: "manual_review",
-    }),
-    [
-      "clawsweeper",
-      "issue-rating: 🧂 unranked krab",
-      "clawsweeper:no-new-fix-pr",
-      "clawsweeper:needs-maintainer-review",
-    ],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["clawsweeper"], {
-      type: "issue",
-      workStatus: "manual_review",
-    }),
-    [
-      "clawsweeper",
-      "issue-rating: 🧂 unranked krab",
-      "clawsweeper:no-new-fix-pr",
-      "clawsweeper:needs-maintainer-review",
-    ],
-  );
-});
-
-test("ClawSweeper issue advisory labels protect queueable issues from stale automation", () => {
-  const queueableLabels = issueAdvisoryLabelsForTest(["bug", "stale"], {
-    type: "issue",
-    workCandidate: "queue_fix_pr",
-    workStatus: "candidate",
-    workConfidence: "high",
-    hasWorkShape: true,
-  });
-
-  assert.equal(queueableLabels.includes("stale"), false);
-  assert.equal(queueableLabels.includes("no-stale"), true);
-  assert.equal(queueableLabels.includes("clawsweeper:queueable-fix"), true);
-  assert.equal(queueableLabels.includes("clawsweeper:fix-shape-clear"), true);
-
-  const alreadyProtectedLabels = issueAdvisoryLabelsForTest(["bug", "no-stale"], {
-    type: "issue",
-    workCandidate: "queue_fix_pr",
-    workStatus: "candidate",
-    workConfidence: "high",
-  });
-
-  assert.equal(alreadyProtectedLabels.includes("stale"), false);
-  assert.equal(alreadyProtectedLabels.filter((label) => label === "no-stale").length, 1);
-  assert.equal(alreadyProtectedLabels.includes("clawsweeper:queueable-fix"), true);
-});
-
-test("ClawSweeper issue advisory labels do not stale-proof non-queueable issues", () => {
-  const lowerConfidenceLabels = issueAdvisoryLabelsForTest(["bug", "stale"], {
-    type: "issue",
-    workCandidate: "queue_fix_pr",
-    workStatus: "candidate",
-    workConfidence: "medium",
-    hasWorkShape: true,
-  });
-
-  assert.equal(lowerConfidenceLabels.includes("stale"), true);
-  assert.equal(lowerConfidenceLabels.includes("no-stale"), false);
-  assert.equal(lowerConfidenceLabels.includes("clawsweeper:queueable-fix"), false);
-
-  const manualReviewLabels = issueAdvisoryLabelsForTest(["bug", "stale"], {
-    type: "issue",
-    workCandidate: "manual_review",
-    workConfidence: "high",
-    hasWorkShape: true,
-  });
-
-  assert.equal(manualReviewLabels.includes("stale"), true);
-  assert.equal(manualReviewLabels.includes("no-stale"), false);
-  assert.equal(manualReviewLabels.includes("clawsweeper:queueable-fix"), false);
-  assert.equal(manualReviewLabels.includes("clawsweeper:fix-shape-clear"), true);
-  assert.equal(manualReviewLabels.includes("clawsweeper:needs-maintainer-review"), true);
-
-  const demotedQueueableLabels = issueAdvisoryLabelsForTest(
-    ["bug", "no-stale", "clawsweeper:queueable-fix"],
-    {
-      type: "issue",
-      workCandidate: "queue_fix_pr",
-      workStatus: "candidate",
-      workConfidence: "medium",
-    },
-  );
-
-  assert.equal(demotedQueueableLabels.includes("no-stale"), false);
-  assert.equal(demotedQueueableLabels.includes("clawsweeper:queueable-fix"), false);
-
-  const manuallyProtectedLabels = issueAdvisoryLabelsForTest(["bug", "no-stale"], {
-    type: "issue",
-    workCandidate: "manual_review",
-  });
-
-  assert.equal(manuallyProtectedLabels.includes("no-stale"), true);
-  assert.equal(manuallyProtectedLabels.includes("clawsweeper:needs-maintainer-review"), true);
-
-  const pullRequestLabels = issueAdvisoryLabelsForTest(["bug", "stale"], {
-    type: "pull_request",
-    workCandidate: "queue_fix_pr",
-    workStatus: "candidate",
-    workConfidence: "high",
-    hasWorkShape: true,
-  });
-
-  assert.deepEqual(pullRequestLabels, ["bug", "stale"]);
-});
-
-test("ClawSweeper issue advisory labels expose linked PR and human decision blockers", () => {
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      hasOpenLinkedPullRequest: true,
-    }),
-    [
-      "bug",
-      "issue-rating: 🧂 unranked krab",
-      "clawsweeper:linked-pr-open",
-      "clawsweeper:no-new-fix-pr",
-    ],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      requiresProductDecision: true,
-    }),
-    [
-      "bug",
-      "issue-rating: 🧂 unranked krab",
-      "clawsweeper:no-new-fix-pr",
-      "clawsweeper:needs-product-decision",
-    ],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      securityReviewStatus: "needs_attention",
-    }),
-    [
-      "bug",
-      "issue-rating: 🧂 unranked krab",
-      "clawsweeper:no-new-fix-pr",
-      "clawsweeper:needs-security-review",
-    ],
-  );
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "issue",
-      itemCategory: "security",
-    }),
-    [
-      "bug",
-      "issue-rating: 🧂 unranked krab",
-      "clawsweeper:no-new-fix-pr",
-      "clawsweeper:needs-security-review",
-    ],
-  );
-});
-
-test("ClawSweeper issue advisory labels remove stale owned labels and preserve other labels", () => {
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(
-      [
-        "bug",
-        "clawsweeper:source-repro",
-        "clawsweeper:not-repro-on-main",
-        "clawsweeper:needs-live-repro",
-        "clawsweeper:needs-info",
-        "clawsweeper:linked-pr-open",
-        "clawsweeper:no-new-fix-pr",
-        "clawsweeper:queueable-fix",
-        "clawsweeper:fix-shape-clear",
-        "clawsweeper:needs-product-decision",
-        "clawsweeper:needs-security-review",
-        "issue-rating: 🦞 diamond lobster",
-        "issue-rating: 🌊 off-meta tidepool",
-        "clawsweeper:autofix",
-        "clawsweeper:automerge",
-        "clawsweeper:human-review",
-        "clawsweeper:merge-ready",
-        "proof: sufficient",
-        "mantis: telegram-visible-proof",
-      ],
-      {
-        type: "issue",
-        reproductionStatus: "reproduced",
-        reproductionConfidence: "high",
-      },
-    ),
-    [
-      "bug",
-      "clawsweeper:autofix",
-      "clawsweeper:automerge",
-      "clawsweeper:human-review",
-      "clawsweeper:merge-ready",
-      "proof: sufficient",
-      "mantis: telegram-visible-proof",
-      "issue-rating: 🦀 challenger crab",
-      "clawsweeper:current-main-repro",
-    ],
-  );
-});
-
-test("ClawSweeper issue advisory labels do not apply to pull requests", () => {
-  assert.deepEqual(
-    issueAdvisoryLabelsForTest(["bug"], {
-      type: "pull_request",
-      reproductionStatus: "reproduced",
-      reproductionConfidence: "high",
-      workCandidate: "queue_fix_pr",
-      workStatus: "candidate",
-      workConfidence: "high",
-      hasOpenLinkedPullRequest: true,
-      requiresProductDecision: true,
-      securityReviewStatus: "needs_attention",
-      hasWorkShape: true,
-    }),
-    ["bug"],
   );
 });
